+++ database/migrations/2023_12_24_084503_create_allowance_by_projects_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('employee_id')->constrained()->onDelete('cascade');
             $table->foreignId('project_id')->constrained()->onDelete('cascade');
             $table->string('allowanceName')->nullable();
-            $table->decimal('amount', 15, 2)->nullable();
+            $table->unsignedInteger('amount')->nullable();
             $table->timestamps();
