+++ database/migrations/2023_12_26_090627_create_allowance_by_others_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('employee_id')->constrained();
             $table->string('allowanceName')->nullable();
-            $table->decimal('amount', 15, 2)->nullable();
+            $table->unsignedInteger('amount')->nullable();
             $table->timestamps();
         });
     }
